Improve the algorithm so that vertices outside of a loop are cached.
This tweak to the algorithm enables it to cache more vertices. Any
vertex that points to (but isn't included in) a loop can now be
cached. There is a special case for an "identity" edge where the
source and destination vertices are the same: this is not counted as a
loop.