A diamond path leading to a loop did not cache what it should have. (#1)
* A diamond path leading to a loop did not cache what it should have.

Start with a graph of the form "a -> b -> c -> b; a -> d -> b;". Here
we have a loop formed by "b->c->b" which leads directly from
a ("a->b") and indirectly via d ("a->d->b"). Previously the algorithm
would consider "d" to be later/deeper than "b" and thus consider it to
be inside the loop, which is clearly wrong.

The change corrects that by deleting entries in the collection of
visited nodes as it back-tracks after having reached the end of each
path. Now the number of entries in that container is always the depth
of the traversal.

Add additional unit tests for this and a couple of other cases.

* Upgrade Linux builds to bionic/libc++-8-dev.

* Can't use the std::get<type> form.

When building in FNV1 mode, the types contained in the
vertex_hash_impl() result tuple are <std::size_t,
std::uint64_t>. These may, of course, be exactly the same.

* Add the loop-within-a-loop test.

* Assert that the 'visited' container is empty before returning.

* Improve naming consistency in the DoubleLoop test.